Example: remove `as any` assertion
In this example, we can remove the need for a type assertion to any;
which is significant foot-gun.
We start by making cleaners for every static type.
If the static type isn't derived from the cleaner via ReturnType
generic, then we must include the static type in the type variable to
asObject.
Then we use those cleaners throughout the cleaner definitions.

Having duplicate definitions for the static type and cleaner is useful
when you need optional fields.
But it would be DRYer to replace the static definition with an alias to the
ReturnType of the cleaner where possible.
If not possible, it should be required that each cleaner includes the
type variable to couple the two definitions.